+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * Support for Python 3.8 has been removed.
   [(#49)](https://github.com/PennyLaneAI/pennylane-honeywell/pull/49)
 
+* The old return type system has been removed.
+  [(#50)](https://github.com/PennyLaneAI/pennylane-honeywell/pull/50)
+
 ### Improvements
 
 ### Documentation